+++ app/src/main/kotlin/org/virtuslab/bazelsteward/app/App.kt
@@ -6,7 +6,7 @@ import org.virtuslab.bazelsteward.bazel.BuildFileSearch
 import org.virtuslab.bazelsteward.bazel.FileUpdateSearch
 import org.virtuslab.bazelsteward.bazel.UpdateLogic
 import org.virtuslab.bazelsteward.core.Workspace
-import org.virtuslab.bazelsteward.maven.MavenDependencyExtractor
+import org.virtuslab.bazelsteward.maven.MavenDataExtractor
 import org.virtuslab.bazelsteward.maven.MavenRepository
 import kotlin.io.path.Path
 
@@ -17,8 +17,8 @@ class App {
       val workspace = Workspace(Path(args[1]))
       val definitions = BuildFileSearch(workspace).buildDefinitions
       val result: String = runBlocking {
-        val currentDependencies = MavenDependencyExtractor(workspace).extract()
-        val availableVersions = MavenRepository().findVersions(currentDependencies)
+        val mavenData = MavenDataExtractor(workspace).extract()
+        val availableVersions = MavenRepository().findVersions(mavenData)
         val updateSuggestions =
           availableVersions.map { UpdateLogic().selectUpdate(it.key, it.value) }.flattenOption()
         val changeSuggestions = FileUpdateSearch(definitions).searchBuildFiles(updateSuggestions)
