+++ maven/src/main/kotlin/org/virtuslab/bazelsteward/maven/MavenData.kt
@@ -1,0 +1,6 @@
+package org.virtuslab.bazelsteward.maven
+
+data class MavenData (
+  val repositories: List<String>,
+  val dependencies: List<MavenCoordinates>
+)
